+++ js-miniapp-sdk/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v1.2.0 (In progress)
 
-- Added support for requesting show ads from the host application to allow displaying Interstitial or Rewarded ads [See here](README.MD#4-Show-Ads)
+- Added support for requesting show ads from the host application to allow displaying Interstitial and Rewarded ads [See here](README.MD#4-Show-Ads)
 
 
 ### v1.1.0 (2020-7-21)