+++ js-miniapp-sdk/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added support for requesting show ads from the host application to allow displaying Interstitial and Rewarded ads [See here](README.MD#4-Show-Ads)
 
+- Added support for requesting load ads from the host application to allow loading Interstitial and Rewarded ads [See here](README.MD#5-Load-Ads)
+
 
 ### v1.1.0 (2020-7-21)