+++ js-miniapp-sdk/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### 1.x.x (In progress)
 
-- **Feature:** Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.MD#4-Show-Ads).
-- **Feature:** Added support for requesting User Name and Profle Photo from Mini app to Host application [See here](README.MD#5-Requesting-User-details).
-- **Feature:** Set and lock device screen orientation. [See here](README.MD#6-Set-screen-orientation).
+- **Feature:** Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.md#4-Show-Ads).
+- **Feature:** Added support for requesting User Name and Profle Photo from Mini app to Host application [See here](README.md#5-Requesting-User-details).
+- **Feature:** Set and lock device screen orientation. [See here](README.md#6-Set-screen-orientation).
 
 ### v1.2.0 (2020-10-02)