+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+### In Progress (2020-XX-XX)
+- **Feature:** Added `rakuten.miniapp.device.location` custom permission.
+
 ### 1.5.0 (2020-11-13)
 
 - **Feature:** Added `MiniApp.getAccessToken` for retrieving an access token. [See here](README.md#6-Get-access-token).