+++ js-miniapp-sdk/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### In Progress (2020-XX-XX)
 - **Feature:** Added `CustomPermissionName.LOCATION`.
-- **Change:** Updated `requestLocationPermission()`` to `requestLocationPermission(permissionDescription?: string)`. From now `requestLocationPermission` will request both custom and device permission respectively.
+- **Change:** Updated `requestLocationPermission()` to `requestLocationPermission(permissionDescription?: string)`. From now `requestLocationPermission` will request both custom and device permission respectively.
 
 ### 1.5.0 (2020-11-13)